+++ packages/dashboard/src/components/TokenIcon/TokenIconSpecialIconList.json
