+++ packages/maskbook/src/extension/options-page/DashboardComponents/TokenIcon.tsx
@@ -9,7 +9,7 @@ import {
 import { Avatar, AvatarProps, makeStyles, Theme } from '@material-ui/core'
 import { useStylesExtends } from '../../../components/custom-ui-helper'
 import { useImageFailover } from '../../../utils'
-import SPECIAL_ICON_LIST from './TokenIconSpeialIconList.json'
+import SPECIAL_ICON_LIST from './TokenIconSpecialIconList.json'
 
 //#region fix icon image
 function resolveTokenIconURLs(address: string, baseURIs: string[], chainId: ChainId, logoURI?: string) {
