+++ packages/maskbook/src/extension/options-page/DashboardComponents/TokenIconSpecialIconList.json
